+++ packages/sqlite_async/example/web/worker.dart
@@ -1,4 +1,4 @@
-import 'package:sqlite_async/sqlite3_web.dart';
+import 'package:sqlite3_web/sqlite3_web.dart';
 import 'package:sqlite_async/sqlite3_web_worker.dart';
 
 void main() {