+++ packages/sqlite_async/lib/native.dart
@@ -1,0 +1,1 @@
+export 'src/native/native_sqlite_open_factory.dart';