+++ packages/sqlite_async/lib/sqlite3_wasm.dart
@@ -1,5 +1,0 @@
-/// Re-exports [sqlite3 WASM](https://pub.dev/packages/sqlite3) to expose sqlite3 without
-/// adding it as a direct dependency.
-library;
-
-export 'package:sqlite3/wasm.dart';