+++ packages/sqlite_async/lib/src/common/abstract_open_factory.dart
@@ -1,48 +1,78 @@
-import 'dart:async';
-import 'package:meta/meta.dart';
+/// @docImport '../native/native_sqlite_open_factory.dart';
+/// @docImport '../web/web_sqlite_open_factory.dart';
+library;
 
-import 'package:sqlite_async/sqlite3_common.dart' as sqlite;
-import 'package:sqlite_async/src/common/mutex.dart';
-import 'package:sqlite_async/src/sqlite_connection.dart';
-import 'package:sqlite_async/src/sqlite_options.dart';
-import 'package:sqlite_async/src/update_notification.dart';
+import 'package:meta/meta.dart';
+import 'package:sqlite3/common.dart' as sqlite;
+
+import '../impl/platform.dart' as platform;
+
+import '../sqlite_options.dart';
 
 /// Factory to create new SQLite database connections.
 ///
 /// Since connections are opened in dedicated background isolates, this class
 /// must be safe to pass to different isolates.
-abstract class SqliteOpenFactory<Database extends sqlite.CommonDatabase> {
-  String get path;
+///
+/// How databases are opened is platform specific. For this reason, this class
+/// can't be extended directly. To customize how databases are opened across
+/// platforms, prefer using [SqliteOptions]. If that class doesn't provide the
+/// degree of customization you need, you can also subclass platform-specific
+/// connection factory implementations:
+///
+///  - On native platforms, extend [NativeSqliteOpenFactory].
+///  - When compiling for the web, extend [WebSqliteOpenFactory].
+sealed class SqliteOpenFactory {
+  final String path;
 
-  /// Opens a direct connection to the SQLite database
-  Database open(SqliteOpenOptions options);
+  SqliteOpenFactory._({required this.path});
 
-  /// Opens an asynchronous [SqliteConnection]
-  FutureOr<SqliteConnection> openConnection(SqliteOpenOptions options);
+  /// Creates a default open factory opening databases at the given path with
+  /// specified options.
+  ///
+  /// This will return a [NativeSqliteOpenFactory] on native platforms and a
+  /// [WebSqliteOpenFactory] when compiling for the web.
+  factory SqliteOpenFactory({
+    required String path,
+    SqliteOptions options = const SqliteOptions(),
+  }) {
+    return platform.createDefaultOpenFactory(path, options);
+  }
+
+  /// Pragma statements to run on newly opened connections to configure them.
+  List<String> pragmaStatements(SqliteOpenOptions options);
 }
 
-class SqliteOpenOptions {
+/// The superclass for all connection factories.
+///
+/// By keeping this class internal, we can safely assert that all connection
+/// factories on native and web platforms extend [NativeSqliteOpenFactory] and
+/// [WebSqliteOpenFactory], respectively.
+@internal
+abstract base class InternalOpenFactory extends SqliteOpenFactory {
+  final SqliteOptions sqliteOptions;
+
+  InternalOpenFactory({
+    required super.path,
+    this.sqliteOptions = const SqliteOptions(),
+  }) : super._();
+}
+
+final class SqliteOpenOptions {
   /// Whether this is the primary write connection for the database.
   final bool primaryConnection;
 
   /// Whether this connection is read-only.
   final bool readOnly;
 
-  /// Mutex to use in [SqliteConnection]s
-  final Mutex? mutex;
-
   /// Name used in debug logs
   final String? debugName;
 
-  /// Stream of external update notifications
-  final Stream<UpdateNotification>? updates;
-
-  const SqliteOpenOptions(
-      {required this.primaryConnection,
-      required this.readOnly,
-      this.mutex,
-      this.debugName,
-      this.updates});
+  const SqliteOpenOptions({
+    required this.primaryConnection,
+    required this.readOnly,
+    this.debugName,
+  });
 
   sqlite.OpenMode get openMode {
     if (primaryConnection) {
@@ -53,61 +83,4 @@ class SqliteOpenOptions {
       return sqlite.OpenMode.readWrite;
     }
   }
-}
-
-/// The default database factory.
-///
-/// This takes care of opening the database, and running PRAGMA statements
-/// to configure the connection.
-///
-/// Override the [open] method to customize the process.
-abstract class AbstractDefaultSqliteOpenFactory<
-        Database extends sqlite.CommonDatabase>
-    implements SqliteOpenFactory<Database> {
-  @override
-  final String path;
-  final SqliteOptions sqliteOptions;
-
-  const AbstractDefaultSqliteOpenFactory(
-      {required this.path,
-      this.sqliteOptions = const SqliteOptions.defaults()});
-
-  List<String> pragmaStatements(SqliteOpenOptions options);
-
-  @protected
-
-  /// Opens a direct connection to a SQLite database connection
-  Database openDB(SqliteOpenOptions options);
-
-  @override
-
-  /// Opens a direct connection to a SQLite database connection
-  /// and executes setup pragma statements to initialize the DB
-  Database open(SqliteOpenOptions options) {
-    var db = openDB(options);
-
-    // Pragma statements don't have the same BUSY_TIMEOUT behavior as normal statements.
-    // We add a manual retry loop for those.
-    for (var statement in pragmaStatements(options)) {
-      for (var tries = 0; tries < 30; tries++) {
-        try {
-          db.execute(statement);
-          break;
-        } on sqlite.SqliteException catch (e) {
-          if (e.resultCode == sqlite.SqlError.SQLITE_BUSY && tries < 29) {
-            continue;
-          } else {
-            rethrow;
-          }
-        }
-      }
-    }
-    return db;
-  }
-
-  @override
-
-  /// Opens an asynchronous [SqliteConnection] to a SQLite database
-  /// and executes setup pragma statements to initialize the DB
-  FutureOr<SqliteConnection> openConnection(SqliteOpenOptions options);
 }