+++ packages/sqlite_async/CHANGELOG.md
@@ -5,6 +5,7 @@
 - __Breaking__: Rewrite the native connection pool implementation.
   - Remove isolate connection factories. Simply open the same database on another isolate, it's safe to do so now.
   - It is no longer possible to register user-defined functions in Dart. Extensions providing functions in native code can still be used.
+- __Breaking__: Remove the `maxReaders` parameter on `SqliteDatabase`. Set that parameter on `SqliteOptions` instead.
 
 ## 0.13.1